Delete SampleView classes
