+++ src/main/java/examples/UseContinuousSample.java
@@ -1,14 +1,14 @@
 package examples;
 
-import datastructs.ContinuousSampleView;
+import datastructs.ContinuousSample;
 
-public class UseSampleView
+public class UseContinuousSample
 {
 
 
 	public static void main(String[] args){
 
-		ContinuousSampleView sample = new ContinuousSampleView(10);
+		ContinuousSample sample = new ContinuousSample("TestSample", 10);
 
 		for(int i=0; i < 10; i++){
 			sample.add(new Double(i));
